Move `ReportFile` into its own file (#552)
* Move `ReportFile` into its own file

* fix imports

* remove broken/duplicated "unittest"
